thumbtackbox
I have the thumbtack weapon in the game now but a lot has chnaged since I wasd last in project and I'm a little confused when ti comes to what's putting the weapon in the arm holder. Was also trying to fix some of the anim things. My debug.drawray for the tackbox dont work though and the bullet isn't being instantiated.