UI Update and Player Description Intro
I added a Windows border for the Ammo UI, as for Health it is represented by a large donut using Image filled. An idea I had was that someone can the Radial 360 into the donut so that the donut goes as the player takes damage, this will still follow the 100 health format. As for the player intro, I left the contract  untouched, but made a different paper intro called Player Evaluation Sheet that serves as the intro as to why the player is this way in a comedic way. It really is up to the group on which paper we use, but we can use both either way as an idea I have is to probably add messages at the end of the stage of update the player personality.