Thumbtack Weapon changes + sounds + animation fixes
I finalized functionality of the thumbtack weapon for the presentation tomorrow. I made sure to add little thumbtack models in the box and it shoots thumbtack projectiles. It also has a sound when you shoot it, and so does the printer. There is also a sound for when the elevator appears.